contextAPI design pattern 리팩토링 중
