+++ 송창석/ebay/src/components/Form/FormInput.jsx
@@ -1,16 +1,15 @@
 import React from 'react';
 
 const FormInput = ({name, type, value, changeValuesHandler}) => {
-    
-    return (
-        <input
-            id={name}
-            name={name}
-            type={type}
-            value={value}
-            onChange={changeValuesHandler}
-        />
-    );
+  return (
+    <input
+      id={name}
+      name={name}
+      type={type}
+      value={value}
+      onChange={changeValuesHandler}
+    />
+  );
 };
 
 export default FormInput;
