+++ 송창석/ebay/src/components/Main/Best/BestContainer.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import BestPresenter from './BestPresenter';
+
+const BestContainer = ({apis}) => {
+  const {title, items} = apis
+
+  return (
+    <BestPresenter
+      title={title}
+      items={items}
+    />
+  );
+};
+
+export default BestContainer;
