+++ 송창석/ebay/src/components/Main/Best/index.jsx
@@ -1,0 +1,2 @@
+import BestContainer from './BestContainer';
+export default BestContainer;
