+++ 송창석/ebay/src/components/Main/Exhibition/EventContainer.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import EventPresenter from './EventPresenter';
+
+const EventContainer = ({apis}) => {
+  const {title, items} = apis
+
+  return (
+    <EventPresenter
+      title={title}
+      items={items}
+    />
+  );
+};
+
+export default EventContainer;
