+++ 송창석/ebay/src/components/Main/Exhibition/index.jsx
@@ -1,0 +1,2 @@
+import EventContainer from './EventContainer';
+export default EventContainer;
