+++ 송창석/ebay/src/components/Main/Gnb/GnbContainer.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import GnbPresenter from './GnbPresenter';
+
+
+
+const GnbContainer = ({apis}) => {
+  const {main, side} = apis
+  
+  return (
+    <GnbPresenter
+      main={main}
+      side={side}
+    />
+  );
+};
+
+export default GnbContainer;
