+++ 송창석/ebay/src/components/Main/Gnb/GnbStyles.jsx
