+++ 송창석/ebay/src/components/Main/Gnb/index.jsx
@@ -1,0 +1,2 @@
+import GnbContainer from './GnbContainer';
+export default GnbContainer;
