--- 송창석/ebay/src/components/Main/MainBest/MainBestIndex.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import MainBestContainer from './MainBestContainer';
-
-const MainBestIndex = () => {
-    return (
-        <MainBestContainer/>
-    );
-};
-
-export default MainBestIndex;
