--- 송창석/ebay/src/components/Main/MainEvent/MainEventIndex.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import MainEventContainer from './MainEventContainer';
-
-const MainEventIndex = () => {
-    return (
-        <MainEventContainer/>
-    );
-};
-
-export default MainEventIndex;
