--- 송창석/ebay/src/components/Main/MainGnb/MainGnbIndex.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import MainGnbContainer from './MainGnbContainer';
-
-const MainGnbIndex = () => {
-    return (
-        <MainGnbContainer/>
-    );
-};
-
-export default MainGnbIndex;
