--- 송창석/ebay/src/components/Main/MainProducts/MainProductsIndex.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import MainProductsContainer from './MainProductsContainer';
-
-const MainProductsIndex = () => {
-    return (
-        <MainProductsContainer/>
-    );
-};
-
-export default MainProductsIndex;
