--- 송창석/ebay/src/components/Main/MainStore/MainStoreIndex.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import MainStoreContainer from './MainStoreContainer';
-
-const MainStoreIndex = () => {
-    return (
-        <MainStoreContainer/>
-    );
-};
-
-export default MainStoreIndex;
