+++ 송창석/ebay/src/components/Main/Products/ProductsContainer.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import ProductsPresenter from './ProductsPresenter';
+
+const ProductsContainer = ({apis}) => {
+  const {title, items} = apis
+
+  return (
+    <ProductsPresenter
+      title={title}
+      items={items}
+    />
+  );
+};
+
+export default ProductsContainer;
