+++ 송창석/ebay/src/components/Main/Products/index.jsx
@@ -1,0 +1,2 @@
+import ProductsContainer from './ProductsContainer';
+export default ProductsContainer;
