+++ 송창석/ebay/src/components/Main/Store/StoreContainer.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import StorePresenter from './StorePresenter';
+
+
+
+const StoreContainer = ({apis}) => {
+  const {title, items} = apis
+  
+  return (
+    <StorePresenter
+      title={title}
+      items={items}
+    />
+  );
+};
+
+export default StoreContainer;
