+++ 송창석/ebay/src/components/Main/Store/index.jsx
@@ -1,0 +1,2 @@
+import StoreContainer from './StoreContainer';
+export default StoreContainer;
