+++ 송창석/ebay/src/routes/Join/UserJoinContainer.jsx
@@ -1,34 +1,25 @@
 import React from 'react';
-import { JOIN_OPTS } from '../../constant/constant';
-import FormUtils from '../../utils/FormUtils';
+import useFormStore from '../../store/useFormStore';
 import UserJoinPresenter from './UserJoinPresenter';
 
-const ConstantContext = React.createContext(null)
-const UserJoinContext = React.createContext(null)
-
-export {
-    ConstantContext,
-    UserJoinContext
-}
+export const UserJoinContext = React.createContext(null)
 
 const UserJoinContainer = () => {
-    const UserJoinValues = FormUtils({
-        name: "",
-        email: "",
-        password: "",
-        gender: "",
-        year: "",
-        month: "",
-        day: ""
-    })
+  const UserJoinValues = useFormStore({
+    name: "",
+    email: "",
+    password: "",
+    gender: "",
+    year: "",
+    month: "",
+    day: ""
+  })
 
-    return (
-        <ConstantContext.Provider value={JOIN_OPTS}>
-            <UserJoinContext.Provider value={UserJoinValues}>
-                <UserJoinPresenter/>
-            </UserJoinContext.Provider>
-        </ConstantContext.Provider>
-    );
+  return (
+    <UserJoinContext.Provider value={UserJoinValues}>
+      <UserJoinPresenter/>
+    </UserJoinContext.Provider>
+  );
 };
 
 export default UserJoinContainer;
