--- 송창석/ebay/src/routes/Join/UserJoinIndex.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import UserJoinContainer from './UserJoinContainer';
-
-const UserJoinIndex = () => {
-    return (
-        <UserJoinContainer/>
-    );
-};
-
-export default UserJoinIndex;
