+++ 송창석/ebay/src/routes/Join/index.jsx
@@ -1,0 +1,2 @@
+import UserJoinContainer from './UserJoinContainer';
+export default UserJoinContainer;
