--- 송창석/ebay/src/routes/Login/LoginIndex.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import LoginContainer from './LoginContainer';
-
-const LoginIndex = () => {
-    return (
-        <LoginContainer/>
-    );
-};
-
-export default LoginIndex;
