+++ 송창석/ebay/src/routes/Main/MainContainer.jsx
@@ -1,34 +1,46 @@
 import React from 'react';
 import { MAIN_APIS } from '../../constant/constant';
-import MainUtils from '../../utils/MainUtils';
+import useMainStore from '../../store/useMainStore';
+import { useOnMounted } from '../../utils/useOnMounted';
 import MainPresenter from './MainPresenter';
 
-const ConstantContext = React.createContext(null)
-const ApisContext = React.createContext(null)
-const DispatchContext = React.createContext(null)
-const FetchDataContext = React.createContext(null)
-
-export {
-    ConstantContext,
-    ApisContext,
-    DispatchContext,
-    FetchDataContext
-}
+export const MainContext = React.createContext(null)
 
 const MainContainer = () => {
-    const {apis, dispatchApis, fetchData} = MainUtils()
+  const {mainState, fetchData} = useMainStore({
+    menu: {
+      main: [],
+      side: []
+    },
+    mainStore: {
+      title: "",
+      items: []
+    },
+    mainBest: {
+      title: "",
+      items: []
+    },
+    mainEvent: {
+      title: "",
+      items: []
+    },
+    mainProducts: {
+      title: "",
+      items: []
+    }
+  })
 
-    return (
-        <ConstantContext.Provider value={MAIN_APIS}>
-            <ApisContext.Provider value={apis}>
-                <DispatchContext.Provider value={dispatchApis}>
-                    <FetchDataContext.Provider value={fetchData}>
-                        <MainPresenter apis={apis}/>
-                    </FetchDataContext.Provider>
-                </DispatchContext.Provider>
-            </ApisContext.Provider>
-        </ConstantContext.Provider>
-    )
+  useOnMounted(() => fetchData("DISPATCH_GNB", MAIN_APIS.MENU))
+  useOnMounted(() => fetchData("DISPATCH_CONTENT", MAIN_APIS.STORE))
+  useOnMounted(() => fetchData("DISPATCH_CONTENT", MAIN_APIS.BEST))
+  useOnMounted(() => fetchData("DISPATCH_CONTENT", MAIN_APIS.EVENT))
+  useOnMounted(() => fetchData("DISPATCH_CONTENT", MAIN_APIS.PRODUCTS))
+
+  return (
+    <MainContext.Provider value={mainState}>
+      <MainPresenter />
+    </MainContext.Provider>
+  )
 }
 
 export default MainContainer;
