--- 송창석/ebay/src/routes/Main/MainIndex.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import MainContainer from './MainContainer';
-
-const MainIndex = () => {
-    return (
-        <MainContainer/>
-    );
-};
-
-export default MainIndex;
