+++ 송창석/ebay/src/routes/Main/MainPresenter.jsx
@@ -1,22 +1,25 @@
 import React from 'react';
-import MainBestIndex from '../../components/Main/MainBest/MainBestIndex';
-import MainEventIndex from '../../components/Main/MainEvent/MainEventIndex';
-import MainGnbIndex from '../../components/Main/MainGnb/MainGnbIndex';
-import MainProductsIndex from '../../components/Main/MainProducts/MainProductsIndex';
-import MainStoreIndex from '../../components/Main/MainStore/MainStoreIndex';
+import Gnb from '../../components/Main/Gnb';
+import Store from '../../components/Main/Store';
+import Best from '../../components/Main/Best';
+import Event from '../../components/Main/Exhibition';
+import Products from '../../components/Main/Products';
 import { MainWrapStyles } from './MainContStyles';
+import { useContext } from 'react';
+import { MainContext } from './MainContainer';
 
 const MainPresenter = () => {
+  const {menu, mainStore, mainBest, mainEvent, mainProducts} = useContext(MainContext)
     
-    return (
-        <MainWrapStyles>
-            <MainGnbIndex/>
-            <MainStoreIndex/>
-            <MainBestIndex/>
-            <MainEventIndex/>
-            <MainProductsIndex/>
-        </MainWrapStyles>
-    );
+  return (
+    <MainWrapStyles>
+      <Gnb apis={menu}/>
+      <Store apis={mainStore}/>
+      <Best apis={mainBest}/>
+      <Event apis={mainEvent}/>
+      <Products apis={mainProducts}/>
+    </MainWrapStyles>
+  );
 };
 
 export default MainPresenter;
