+++ 송창석/ebay/src/routes/Main/index.jsx
@@ -1,0 +1,2 @@
+import MainContainer from './MainContainer';
+export default MainContainer;
