+++ 송창석/ebay/src/utils/fetch.js
@@ -1,6 +1,6 @@
 const toJson = res => res.json()
 
 export const fetchJson = (name) => {
-    const api = fetch(`https://baekcode.github.io/APIs/${name}.json`).then(toJson)
-    return api;
+  const api = fetch(`https://baekcode.github.io/APIs/${name}.json`).then(toJson)
+  return api;
 }
