+++ 송창석/ebay/src/utils/useOnMounted.js
@@ -1,0 +1,5 @@
+import { useEffect } from "react"
+
+export const useOnMounted = (effect) => {
+  useEffect(effect, [])
+}
